homenu

+++ app/news/views/lists.php
@@ -4,6 +4,7 @@ $msg_type = getFlashData('msg_type');
 if (!empty($_GET['keyword'])) {
     $keyword = $_GET['keyword'];
 }
+$limit=55;
 ?>
 <div class="container-fluid">
     <a href="?module=news&action=add"><button class="btn btn-success">Thêm tin tức tin tức mới <i
@@ -38,27 +39,27 @@ if (!empty($_GET['keyword'])) {
             </tr>
         </thead>
         <tbody>
-            <?php if (!empty($data['news'])) :
-                foreach ($data['news'] as $key => $item) : ?>
+        <?php
+            $uploadDir = '../uploads/news/'; 
+            if(!empty($data['news'])):
+                foreach($data['news'] as $key => $item): ?>
             <tr>
-                <td><?php echo $key + 1 ?></td>
-                <td><?php echo $item['thumbnail'] ?></td>
+                <td><?php echo $key+1 ?></td>
+                <td>
+                            <?php
+                            if (is_file($uploadDir . $item['thumbnail'])) {
+                                echo "<img src='" . $uploadDir . $item['thumbnail'] . "' height='80' width='80' >";
+                            } else {
+                                echo "Không có ảnh";
+                            }
+                            ?>
+                        </td>
                 <td><?php echo $item['title'] ?></td>
-                <td><?php echo $item['content'] ?></td>
-                <td><a href="?module=news&action=edit&id=<?php echo $item['id'] ?>"><button class="btn btn-warning"><i
-                                class="fa fa-edit"></i></button></a></td>
-                <td><a href="?module=news&action=delete&id=<?php echo $item['id'] ?>"
-                        onclick="return confirm('Bạn có chắc chắn muốn xóa?')"><button class="btn btn-danger"><i
-                                class="fa fa-trash"></i></button></a></td>
-            </tr>
-            <?php endforeach;
-            else:
-                ?>
-            <td colspan="6">
-                <div class="alert alert-warning text-center">Thông tin bạn tìm kiếm không có dữ liệu</div>
-            </td>
-            <?php
-            endif; ?>
+                <td><?php echo substr($item['content'],0,$limit) ?></td>
+                <td><a href="?module=news&action=edit&id=<?php echo $item['id'] ?>"><button class="btn btn-warning"><i class="fa fa-edit"></i></button></a></td>
+                <td><a href="?module=news&action=delete&id=<?php echo $item['id'] ?>" onclick="return confirm('Bạn có chắc chắn muốn xóa?')"><button class="btn btn-danger"><i class="fa fa-trash"></i></button></a></td>
+                </tr>
+            <?php endforeach; endif; ?>
         </tbody>
     </table>
 </div>
